Extract TerminalInputCoordinator from TerminalInputTextView
DECOMPOSITION-PLAN 2c input split: the armed/sticky modifier machine and
every committed-text / backspace / accessory-tap byte translation move into
a @MainActor @Observable TerminalInputCoordinator in CmuxMobileTerminalKit,
tested with 14 Swift Testing cases. The UITextView is now a dumb first
responder that forwards events and dispatches the returned emissions.

Resolutions are returned values rather than the plan's AsyncStream<Data>:
this is the per-keystroke typing-latency path, and the prefer-values rule
beats an extra main-queue hop per keystroke (probe re-measured: processed
p50 10.0ms, 60Hz cadence lock, no regression vs baseline).

TerminalInputAccessoryAction moves to CmuxMobileTerminalKit (pure value the
coordinator consumes; members publicized with DocC) and the hardware-key
resolver gets its own file, shrinking GhosttySurfaceView.swift further.
